tweak how to find ns-3 libraries
ns3 library names change according to version and profile.
DCE looks for libraries with the same profile as its own (eg,
"debug"/"optimized").
Previous to this patch, DCE tries to match the names against a list of
supported version.
The main issue at the moment is that libraries for different modules can
be loaded with different versions.
With this patch, we check the version of the core library and make sure
to check modules against that version specifically.
If there are several `core` libraries available than DCE errors and the
user should delete the additionnal versions.
As long as one can install different versions in the same folder, it is
possible to mess up things (while nix forbids to change an installed library).

In order to solve https://github.com/direct-code-execution/ns-3-dce/pull/99,
we would need ns-3 to rename the `-dev` version into `X-dev` where X is
the version of the next release.