Fix bug in dce_vasprintf
The function dce_vasprintf calls to the system vasprintf, using a ret variable to ask for memory and copy the string that is returned by vasprintf. ret is the amount of bytes in the string without considering the termination character \0. For that reason there is a bug here and ret should be ret+1.